+++ src/llm-providers/openai/OpenAiAssistantExecutionTools.ts
@@ -7,13 +7,15 @@ import type { ChatPromptResult } from '../../execution/PromptResult';
 import { UNCERTAIN_USAGE } from '../../execution/utils/usage-constants';
 import type { ModelRequirements } from '../../types/ModelRequirements';
 import type { Prompt } from '../../types/Prompt';
-import type { string_date_iso8601 } from '../../types/typeAliases';
-import type { string_markdown } from '../../types/typeAliases';
-import type { string_markdown_text } from '../../types/typeAliases';
-import type { string_title } from '../../types/typeAliases';
-import type { string_token } from '../../types/typeAliases';
+import type {
+    string_date_iso8601,
+    string_markdown,
+    string_markdown_text,
+    string_title,
+    string_token,
+} from '../../types/typeAliases';
 import { getCurrentIsoDate } from '../../utils/getCurrentIsoDate';
-import { replaceParameters } from '../../utils/replaceParameters';
+import { replaceParameters } from '../../utils/parameters/replaceParameters';
 import { $asDeeplyFrozenSerializableJson } from '../../utils/serialization/$asDeeplyFrozenSerializableJson';
 import type { OpenAiAssistantExecutionToolsOptions } from './OpenAiAssistantExecutionToolsOptions';
 import { OpenAiExecutionTools } from './OpenAiExecutionTools';
@@ -33,8 +35,9 @@ export class OpenAiAssistantExecutionTools extends OpenAiExecutionTools implemen
      *
      * @param options which are relevant are directly passed to the OpenAI client
      */
-    public constructor(options: OpenAiAssistantExecutionToolsOptions = {}) {
+    public constructor(options: OpenAiAssistantExecutionToolsOptions) {
         super(options);
+        this.assistantId = options.assistantId;
     }
 
     public get title(): string_title & string_markdown_text {